Boton importar Csv v1.3.3

+++ src/main/java/ucv/codelab/repository/BaseRepository.java
@@ -107,7 +107,7 @@ public abstract class BaseRepository<T> {
         String sql = "SELECT * FROM " + getTableName() + " WHERE " + getIdColumnName() + " = ?";
 
         try (PreparedStatement stmt = connection.prepareStatement(sql)) {
-            stmt.setInt(1, (Integer) id);
+            stmt.setInt(1, id);
 
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
